refactor: DashboardRepository 구조 변경
