+++ src/main/java/org/guzzing/studayserver/domain/dashboard/repository/DashboardRepository.java
@@ -3,17 +3,13 @@ package org.guzzing.studayserver.domain.dashboard.repository;
 import java.util.List;
 import java.util.Optional;
 import org.guzzing.studayserver.domain.dashboard.model.Dashboard;
-import org.guzzing.studayserver.global.exception.DashboardException;
 import org.springframework.data.jpa.repository.Query;
 
 public interface DashboardRepository {
 
     Dashboard save(final Dashboard dashboard);
 
-    default Dashboard findDashboardById(final Long dashboardId) {
-        return this.findById(dashboardId)
-                .orElseThrow(() -> new DashboardException("존재하지 않는 대시보드입니다."));
-    }
+    Dashboard findDashboardById(final Long dashboardId);
 
     Optional<Dashboard> findById(final Long dashboardId);
